Created "Totem" and "Totem Manager" Classes
These classes instanciate rabbit totem objects and fox totem objects
using the List<t> type to store them

+++ Unity With Zach 1 - 2D Project/Assets/FoxPop.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class FoxPop : MonoBehaviour {
 
-    public double fCount;
+    public int fCount;
     public double fEfficiency;
-    public double fRabbitsEatenToGrow;
+    public int fRabbitsEatenToGrow;
 
     // Use this for initialization
     void Start () {
